// Microservice.Web/src/app/model/RatingModel.ts
import { Component } from '@angular/core';

export class RatingModel {
    Rater: number;
    Ratee: number;
    TypeId: number;
    RateId: number;
    GoalId: number;
    GoalType: number;
    Rate: number;
    Comment: string;
}

export class RatingUpdateModel {

    // Id: number;
    // GoalId: number;
    // GoalType: number;
    // Rate: number;
    // Comment: string;
    // CreatedOn: Date;
    // CreatedBy: number;
    // UpdatedOn: Date;
    // UpdatedBy: number;

}